refactor(utils): names, files, commends, code
