+++ package/configs/jest/setup.ts
@@ -1,6 +1,6 @@
 import 'jest-preset-angular';
 import './global-mocks';
 
-import { reset } from '../../src/utils/helpers';
+import { resetPackage } from '../../src/utils/helpers';
 
-afterEach(reset);
+afterEach(resetPackage);
