+++ package/src/utils/link.ts
@@ -1,5 +1,6 @@
 import { Structure } from '../interfaces';
 import { transformPathToState } from './path';
+import { showError } from './helpers';
 
 /**
  * makes adjustments to string link
@@ -7,10 +8,7 @@ import { transformPathToState } from './path';
 export function correctStringLink(value: string): string {
   switch (true) {
     case value === '[Object object]': {
-      // tslint:disable-next-line:max-line-length
-      console.error(
-        `ERROR: incorrect [[${value}]] has been passed into navLink`
-      );
+      showError(`ERROR: incorrect [[${value}]] has been passed into navLink.`);
       return '/';
     }
     case value.includes(','):
@@ -38,7 +36,7 @@ export const getRouteLink = (
   } else if (typeof input === 'object' && input.state) {
     return input.state;
   } else {
-    console.error('ERROR: navLink gets incorrect data format', '\n', input);
+    showError('ERROR: navLink gets incorrect data format', '\n', input);
     return [];
   }
 };
