+++ package/src/utils/path.ts
@@ -1,3 +1,5 @@
+import { DefaultNameOptions } from '../interfaces';
+
 /**
  * removes empty or slash paths
  */
@@ -20,3 +22,50 @@ export const transformPathToState = (
  * checks a string on wildcard path
  */
 export const isWildcard = (path: string) => path === '**';
+
+/**
+ * cuts off unwanted strings
+ */
+function fixPathName(path: string): string {
+  let newPath = '';
+  let index = path.length - 1;
+
+  if (path[index] === '/') {
+    index -= 1;
+  }
+
+  while (index > -1) {
+    const char = path[index];
+    const nextChar = path[index - 1];
+    if (nextChar === '/' || nextChar === ':') {
+      newPath = char + newPath;
+      break;
+    } else if (nextChar === '-' || nextChar === '_') {
+      newPath = char.toUpperCase() + newPath;
+      index -= 1;
+    } else {
+      newPath = char + newPath;
+    }
+
+    index -= 1;
+  }
+
+  return newPath;
+}
+
+/**
+ * sets a route name
+ * based on path and options
+ */
+export function setRouteName(
+  path: string,
+  nameOptions: DefaultNameOptions = {}
+): string {
+  if (path === '') {
+    return nameOptions.root || 'root';
+  } else if (isWildcard(path)) {
+    return nameOptions.wildcard || 'wildcard';
+  } else {
+    return fixPathName(path);
+  }
+}
